Separate files, change export to function (vs obj)
- Reshaper is now called simply with reshaper(), rather than
  reshaper.findShape().
- Extract and traverse operations have been separated out
- Exceptions are now used for error handling, instead of just returning
  null.
