+++ lib/util.js
@@ -17,7 +17,18 @@ function removeFromArray(array, from, to) {
   return array.push.apply(array, rest);
 };
 
+function NotFoundException(data, schema, hint) {
+    this.name = 'NotFoundException';
+    this.stack = (new Error()).stack;
+    this.message = 'Could not find';
+    this.data = data;
+    this.schema = schema;
+    this.hint = hint;
+}
+NotFoundException.prototype = Object.create(Error.prototype);
+NotFoundException.prototype.constructor = NotFoundException;
 
 exports.typeString = typeString;
 exports.typesMatch = typesMatch;
 exports.removeFromArray = removeFromArray;
+exports.NotFoundException = NotFoundException;
