+++ test/basic.js
@@ -39,35 +39,35 @@ describe('basic', function() {
     ];
 
     it('should return array of x values without hinting', function() {
-        var result = reshaper.findShape(data, schema);
+        var result = reshaper(data, schema);
         expect(result).to.eql([12, 2, 9]);
     });
 
     it('should return array of y values with y hint', function() {
-        var result = reshaper.findShape(data, schema, 'y');
+        var result = reshaper(data, schema, 'y');
         expect(result).to.eql([5, 3, 14]);
     });
 
     it('should extract ages from people data', function() {
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql([23, 24]);
     });
 
     it('should extract first names from people data', function() {
         var schema = ['String'];
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql(['Joel', 'Jake']);
     });
 
     it('should extract last names from people data', function() {
         var schema = ['String'];
-        var result = reshaper.findShape(peopleData, schema, 'lastName');
+        var result = reshaper(peopleData, schema, 'lastName');
         expect(result).to.eql(['Auterson', 'Hall']);
     });
 
     it('should handle booleans', function() {
         var schema = ['Boolean'];
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql([true, false]);
     });
 
@@ -76,7 +76,7 @@ describe('basic', function() {
             nums: [1, 2, 3],
             strs: ['a', 'b', 'c']
         };
-        var result = reshaper.findShape(data, ['String']);
+        var result = reshaper(data, ['String']);
         expect(result).to.eql(['a', 'b', 'c']);
     });
 
@@ -92,7 +92,7 @@ describe('basic', function() {
             }
         };
         var schema = {one: 'String', two: 'String'};
-        expect(reshaper.findShape(data, schema)).to.eql({
+        expect(reshaper(data, schema)).to.eql({
             one: 'e',
             two: 'f'
         });
@@ -110,7 +110,7 @@ describe('basic', function() {
             }
         };
         var schema = {one: 'String', two: 'String'};
-        expect(reshaper.findShape(data, schema, {one: 'c'})).to.eql({
+        expect(reshaper(data, schema, {one: 'c'})).to.eql({
             one: 'e',
             two: 'f'
         });
@@ -119,7 +119,7 @@ describe('basic', function() {
 
     it('should extract arrays of simple objects from people data', function() {
         var schema = [{name: 'String', age: 'Number'}];
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql([
             {
                 name: 'Joel',
@@ -146,7 +146,7 @@ describe('basic', function() {
                 }
             }
         ];
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql(peopleData);
     });
 
@@ -155,7 +155,7 @@ describe('basic', function() {
             age: ['Number'],
             height: ['Number']
         };
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql({
             age: [23, 24],
             height: [1.9, 1.85]
@@ -167,7 +167,7 @@ describe('basic', function() {
             name: ['String'],
             lastName: ['String']
         };
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         expect(result).to.eql({
             name: ['Joel', 'Jake'],
             lastName: ['Auterson', 'Hall']
@@ -180,25 +180,25 @@ describe('basic', function() {
             lastName: ['String'],
             middleName: ['String']
         };
-        var result = reshaper.findShape(peopleData, schema);
+        var result = reshaper(peopleData, schema);
         schema = {
             lastName: ['String'],
             middleName: ['String'],
             name: ['String']
         };
-        var result2 = reshaper.findShape(peopleData, schema);
+        var result2 = reshaper(peopleData, schema);
         expect(result2).to.eql(result);
     });
 
     it('should, with no hint, pick the shallowest', function() {
         var schema = ['String'];
-        var result = reshaper.findShape(peopleData, schema, 'firstName');
+        var result = reshaper(peopleData, schema, 'firstName');
         expect(result).to.eql(['Joel', 'Jake']);
         schema = {
             firstName: ['String'],
             lastName: ['String']
         };
-        result = reshaper.findShape(peopleData, schema);
+        result = reshaper(peopleData, schema);
         expect(result).to.eql({
             firstName: ['Joel', 'Jake'],
             lastName: ['Auterson', 'Hall']
@@ -230,7 +230,7 @@ describe('basic', function() {
             width: ['Number'],
             red: ['Number']
         };
-        var result = reshaper.findShape(rectangles, schema);
+        var result = reshaper(rectangles, schema);
         expect(result).to.eql({
             width: [5, 10],
             red: [255, 50]
@@ -246,7 +246,7 @@ describe('basic', function() {
         var schema = {
             colour: ['Number']
         };
-        var result = reshaper.findShape(data, schema);
+        var result = reshaper(data, schema);
         expect(result).to.eql({
             colour: [1, 2, 3]
         });
@@ -262,12 +262,12 @@ describe('basic', function() {
             }
         };
         var schema = ['Number'];
-        var result = reshaper.findShape(data, schema);
+        var result = reshaper(data, schema);
         expect(result).to.eql([1, 2]);
     });
 
     it('should allow hints to be used within objects', function() {
-        var result = reshaper.findShape(peopleData, genericSchema, 'height');
+        var result = reshaper(peopleData, genericSchema, 'height');
         expect(result).to.eql([
             {x: 'Joel', y: 1.9},
             {x: 'Jake', y: 1.85}
@@ -275,7 +275,7 @@ describe('basic', function() {
     });
 
     it('should allow multiple hints to be used', function() {
-        var result = reshaper.findShape(peopleData, genericSchema, ['lastName', 'height']);
+        var result = reshaper(peopleData, genericSchema, ['lastName', 'height']);
         expect(result).to.eql([
             {x: 'Auterson', y: 1.9},
             {x: 'Hall', y: 1.85}
@@ -285,7 +285,7 @@ describe('basic', function() {
     it('should fix on a used key', function() {
         var data = [{a: 1, b: 2}, {b: 3, a: 4}];
         var schema = ['Number'];
-        var result = reshaper.findShape(data, schema);
+        var result = reshaper(data, schema);
         expect(result).to.eql([1, 4]);
     });
 
@@ -294,16 +294,33 @@ describe('basic', function() {
             x: ['String'],
             y: ['String']
         };
-        var result = reshaper.findShape(peopleData, schema, ['lastName', 'middleName']);
+        var result = reshaper(peopleData, schema, ['lastName', 'middleName']);
         expect(result).to.eql({
             x: ['Auterson', 'Hall'],
             y: ['Robert', 'Wild']
         });
-        var result = reshaper.findShape(peopleData, schema, ['middleName', 'lastName']);
+        var result = reshaper(peopleData, schema, ['middleName', 'lastName']);
         expect(result).to.eql({
             y: ['Auterson', 'Hall'],
             x: ['Robert', 'Wild']
         });
+    });
+
+    it('should throw exception if cannot create array from object', function() {
+        var data = {
+            a: 'a',
+            b: 'b'
+        };
+        var schema = ['Number'];
+        expect(function() {
+            reshaper(data, schema)
+        }).to.throw('Could not find');
+    });
+
+    it('should throw exception if cannot create array from primitive', function() {
+        expect(function() {
+            reshaper('a', ['Number'])
+        }).to.throw('Could not find');
     });
 
 });
